Update uri-templates to 0.1.5
