+++ uri-templates/uri-templates-tests-0.1.2.ts
@@ -1,0 +1,16 @@
+/// <reference path="uri-templates-0.1.2.d.ts" />
+
+import utmpl = require('uri-templates');
+import URITemplate = utmpl.URITemplate;
+
+var str: string;
+var u: URITemplate;
+var obj: Object;
+
+u = utmpl(str);
+
+str = u.fillFromObject(obj);
+str = u.fill((key) => {
+	return str;
+});
+obj = u.fromUri(str);
